core build with eCAL::v5::CSubscriber
